Moved the scoring functions to abstractlocus. This should simplify code mantainance.
