+++ Mikado/loci/locus.py
@@ -9,7 +9,6 @@ import collections
 import itertools
 import operator
 from collections import deque
-from sys import version_info
 import pyfaidx
 from ..transcripts.transcript import Transcript
 from ..transcripts.transcriptchecker import TranscriptChecker
@@ -18,10 +17,6 @@ from .sublocus import Sublocus
 from ..parsers.GFF import GffLine
 from ..scales.assigner import Assigner
 from ..utilities import overlap
-if version_info.minor < 5:
-    from sortedcontainers import SortedDict
-else:
-    from collections import OrderedDict as SortedDict
 
 
 class Locus(Sublocus, Abstractlocus):
@@ -68,6 +63,7 @@ class Locus(Sublocus, Abstractlocus):
         self.__id = None
         self.fai = None
         self.json_conf = json_conf
+        self.purge = self.json_conf["pick"]["clustering"]["purge"]
         # if verified_introns is not None:
         #     self.locus_verified_introns = verified_introns
 
@@ -393,25 +389,22 @@ reached the maximum number of isoforms for the locus".format(
             raise TypeError("Invalid configuration of type {0}".format(type(jconf)))
         self.json_conf = jconf
 
-    def get_metrics(self):
-
-        """Quick wrapper to calculate the metrics for all the transcripts."""
-
-        # TODO: Find an intelligent way ot restoring this check
-
-        if self.metrics_calculated is True:
-            return
-
-        # self.logger.info("Calculating the intron tree for %s", self.id)
-        assert len(self._cds_introntree) == len(self.combined_cds_introns)
-
-        for tid in sorted(self.transcripts):
-            self.calculate_metrics(tid)
-
-        self.logger.debug("Finished to calculate the metrics for %s", self.id)
-
-        self.metrics_calculated = True
-        return
+    # def get_metrics(self):
+    #
+    #     """Quick wrapper to calculate the metrics for all the transcripts."""
+    #
+    #     if self.metrics_calculated is True:
+    #         return
+    #
+    #     assert len(self._cds_introntree) == len(self.combined_cds_introns)
+    #
+    #     for tid in sorted(self.transcripts):
+    #         self.calculate_metrics(tid)
+    #
+    #     self.logger.debug("Finished to calculate the metrics for %s", self.id)
+    #
+    #     self.metrics_calculated = True
+    #     return
 
     def calculate_metrics(self, tid: str):
         """
@@ -463,59 +456,16 @@ reached the maximum number of isoforms for the locus".format(
         if self.scores_calculated is True:
             return
 
+        self.scores = dict()
         self.get_metrics()
+
+        super().calculate_scores()
+
         if not hasattr(self, "logger"):
             self.logger = None
             self.logger.setLevel("DEBUG")
         self.logger.debug("Calculating scores for {0}".format(self.id))
 
-        self.scores = dict()
-        for tid in self.transcripts:
-            self.scores[tid] = dict()
-            # Add the score for the transcript source
-            self.scores[tid]["source_score"] = self.transcripts[tid].source_score
-
-        if self.regressor is None:
-            for param in self.json_conf["scoring"]:
-                self._calculate_score(param)
-
-            for tid in self.scores:
-                self.transcripts[tid].scores = self.scores[tid].copy()
-
-            for tid in self.transcripts:
-
-                if tid in self.__orf_doubles:
-                    del self.scores[tid]
-                    continue
-                self.transcripts[tid].score = sum(self.scores[tid].values())
-                self.scores[tid]["score"] = self.transcripts[tid].score
-
-        else:
-            valid_metrics = self.regressor.metrics
-            metric_rows = SortedDict()
-            for tid, transcript in sorted(self.transcripts.items(), key=operator.itemgetter(0)):
-                for param in valid_metrics:
-                    self.scores[tid][param] = "NA"
-                row = []
-                for attr in valid_metrics:
-                    val = getattr(transcript, attr)
-                    if isinstance(val, bool):
-                        if val:
-                            val = 1
-                        else:
-                            val = 0
-                    row.append(val)
-                metric_rows[tid] = row
-            # scores = SortedDict.fromkeys(metric_rows.keys())
-            for pos, score in enumerate(self.regressor.predict(list(metric_rows.values()))):
-                tid = list(metric_rows.keys())[pos]
-                if tid in self.__orf_doubles:
-                    del self.scores[tid]
-                    continue
-                self.scores[tid]["score"] = score
-                self.transcripts[tid].score = score
-
-        self.metric_lines_store = []
         for row in self.prepare_metrics():
             if row["tid"] in self.__orf_doubles:
                 continue
@@ -526,8 +476,8 @@ reached the maximum number of isoforms for the locus".format(
             for partial in self.__orf_doubles[doubled]:
                 if partial in self.transcripts:
                     del self.transcripts[partial]
-
-        self.scores_calculated = True
+                if partial in self.scores:
+                    del self.scores[partial]
 
     def print_scores(self):
         """This method yields dictionary rows that are given to a csv.DictWriter class."""
@@ -542,19 +492,26 @@ reached the maximum number of isoforms for the locus".format(
             row = dict().fromkeys(keys)
             row["tid"] = tid
             row["parent"] = self.id
-            row["score"] = round(self.scores[tid]["score"], 2)
+            if tid in self._not_passing:
+                row["score"] = 0
+            else:
+                row["score"] = round(self.scores[tid]["score"], 2)
             calculate_total = (self.regressor is None)
             for key in score_keys:
                 if calculate_total:
                     assert self.scores[tid][key] != "NA" and self.scores[tid][key] is not None
                     row[key] = round(self.scores[tid][key], 2)
+
             if calculate_total is True:
                 score_sum = sum(row[key] for key in score_keys)
-                #
-                assert round(score_sum, 2) == round(self.scores[tid]["score"], 2), (
-                    score_sum,
-                    self.transcripts[tid].score,
-                    tid)
+                if tid not in self._not_passing and self.scores[tid]["score"] > 0:
+                    assert round(score_sum, 2) == round(self.scores[tid]["score"], 2), (
+                        score_sum,
+                        self.transcripts[tid].score,
+                        tid)
+                else:
+                    assert self.scores[tid]["score"] == 0
+
             yield row
 
     def is_alternative_splicing(self, other):
@@ -658,7 +615,7 @@ reached the maximum number of isoforms for the locus".format(
 
             comm = five_comm.popleft()
             comm = deque(sorted(list(set.difference(set(comm), five_found)),
-                         key=lambda tid: self[tid].start))
+                         key=lambda internal_tid: self[internal_tid].start))
             if len(comm) == 1:
                 continue
             first = comm.popleft()
@@ -688,7 +645,7 @@ reached the maximum number of isoforms for the locus".format(
 
             comm = three_comm.popleft()
             comm = deque(sorted(list(set.difference(set(comm), three_found)),
-                         key=lambda tid: self[tid].end, reverse=True))
+                         key=lambda internal_tid: self[internal_tid].end, reverse=True))
             if len(comm) == 1:
                 continue
             first = comm.popleft()
@@ -709,7 +666,7 @@ reached the maximum number of isoforms for the locus".format(
                     three_found.add(tid)
                 else:
                     continue
-            comm = deque([_ for _ in comm if _ not in three_found ])
+            comm = deque([_ for _ in comm if _ not in three_found])
             if comm:
                 three_comm.appendleft(comm)
 
@@ -889,11 +846,7 @@ def expand_transcript(transcript, new_start, new_end, fai, logger):
                 raise AssertionError(err)
             logger.debug("New ORF: %s", str(orf))
             new_orfs.append(orf)
-        # from ..utilities.log_utils import create_default_logger
-        # transcript.logger = create_default_logger("TEMP")
-        # transcript.logger.setLevel("DEBUG")
         transcript.load_orfs(new_orfs)
-        # transcript.logger.setLevel("WARNING")
 
     # Now finalize again
     transcript.finalize()
