+++ Mikado/tests/test_system_calls.py
@@ -642,7 +642,9 @@ Chr1	foo	exon	19000	20000	.	+	.	gene_id "foo"; transcript_id "foo2.1"""
             with self.subTest(purging=purging):
                 json_conf["pick"]["files"]["loci_out"] = "mikado.purging_{}.loci.gff3".format(purging)
                 json_conf["pick"]["files"]["subloci_out"] = "mikado.purging_{}.subloci.gff3".format(purging)
-                json_conf["pick"]["files"]["log"] = "mikado.purging_{}.log".format(purging)
+                json_conf["pick"]["files"]["log"] = os.path.join(
+                    tempfile.gettempdir(),
+                    "mikado.purging_{}.log".format(purging))
                 json_conf["pick"]["clustering"]["purge"] = purging
                 json_conf["pick"]["scoring_file"] = scoring_file.name
                 json_conf = configurator.check_json(json_conf)
